+++ py/desispec/test/test_spectra.py
@@ -40,7 +40,9 @@ class TestSpectra(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
         """Create specprod directory structure"""
+        cls.origDir = os.getcwd()
         cls.testDir = tempfile.mkdtemp()
+        os.chdir(cls.testDir)
         cls.origEnv = {
             "SPECPROD": None,
             "DESI_SPECTRO_REDUX": None,
@@ -72,8 +74,11 @@ class TestSpectra(unittest.TestCase):
             else:
                 os.environ[e] = cls.origEnv[e]
 
-        if os.path.exists(cls.testDir):
+        #- Remove testdir only if it was created by tempfile.mkdtemp
+        if cls.testDir.startswith(tempfile.gettempdir()) and os.path.exists(cls.testDir):
             shutil.rmtree(cls.testDir)
+
+        os.chdir(cls.origDir)
 
 
     def setUp(self):
@@ -81,6 +86,7 @@ class TestSpectra(unittest.TestCase):
         # warnings.filterwarnings("error", ".*did not parse as fits unit.*")
 
         #- Test data and files to work with
+        os.chdir(self.testDir)
         self.fileio = "test_spectra.fits"
         self.fileappend = "test_spectra_append.fits"
         self.filebuild = "test_spectra_build.fits"
